fix unsigned int typecast in loop iterator

+++ PlaceRouteHierFlow/placer/SeqPair.cpp
@@ -1173,7 +1173,7 @@ std::string SeqPair::getLexIndex(design& des) const {
 bool SeqPair::CheckSymm(design& caseNL) {
   auto logger = spdlog::default_logger()->clone("placer.SeqPair.CheckSymm");
   std::map<int, int> posPosition, negPosition;
-  for (int i = 0; i < posPair.size(); ++i) {
+  for (int i = 0; i < ((int)posPair.size()); ++i) {
     posPosition[posPair[i]] = i;
     negPosition[negPair[i]] = i;
   }
@@ -1181,10 +1181,10 @@ bool SeqPair::CheckSymm(design& caseNL) {
 
     // self symm blocks should be (above/below for vertical axis) or (left/right for horizontal axis)
     // self symm blocks to the (left/right for vertical axis) or (above/below) for horizontal) is a violation
-    for (int i = 0; i < sb.selfsym.size() - 1; ++i) {
+    for (int i = 0; i < ((int)sb.selfsym.size()) - 1; ++i) {
       auto posA = posPosition[sb.selfsym[i].first];
       auto negA = negPosition[sb.selfsym[i].first];
-      for (int j = i + 1; j < sb.selfsym.size(); ++j) {
+      for (int j = i + 1; j < ((int)sb.selfsym.size()); ++j) {
         auto posB = posPosition[sb.selfsym[j].first];
         auto negB = negPosition[sb.selfsym[j].first];
         if (sb.axis_dir == placerDB::V) {
@@ -1198,7 +1198,7 @@ bool SeqPair::CheckSymm(design& caseNL) {
         }
       }
     }
-    for (int i = 0; i < sb.sympair.size(); ++i) {
+    for (int i = 0; i < ((int)sb.sympair.size()); ++i) {
       const auto& sympairi = sb.sympair[i];
       auto posA = posPosition[sympairi.first];
       auto negA = negPosition[sympairi.first];
@@ -1221,7 +1221,7 @@ bool SeqPair::CheckSymm(design& caseNL) {
               return false;
           }
         }
-        for (int j = i+1; j < sb.sympair.size(); ++j) {
+        for (int j = i+1; j < ((int)sb.sympair.size()); ++j) {
           const auto& sympairj = sb.sympair[j];
           auto posC = posPosition[sympairj.first];
           auto negC = negPosition[sympairj.first];
@@ -1266,7 +1266,7 @@ bool SeqPair::CheckSymm(design& caseNL) {
             return false;
           }
         }
-        for (int j = i+1; j < sb.sympair.size(); ++j) {
+        for (int j = i+1; j < ((int)sb.sympair.size()); ++j) {
           const auto& sympairj = sb.sympair[j];
           auto posC = posPosition[sympairj.first];
           auto negC = negPosition[sympairj.first];
@@ -1378,7 +1378,7 @@ bool SeqPair::CheckSymm(design& caseNL) {
         }
       }
     }
-    for (int i = posA + 1; i < posPair.size(); ++i) {
+    for (int i = posA + 1; i < ((int)posPair.size()); ++i) {
       const auto& bi = posPair[i];
       if (bi < caseNL.Blocks.size()) {
         if (negPosition[bi] < negA) {
@@ -1431,11 +1431,11 @@ bool SeqPair::CheckSymm(design& caseNL) {
       for (const auto& au : caseNL.Align_blocks) {
         if (au.line == 0) {
           if ((horiz && au.horizon == 1) || (!horiz && au.horizon == 0)) {
-            for (int i = 0; i < au.blocks.size() - 1; ++i) {
+            for (int i = 0; i < ((int)au.blocks.size()) - 1; ++i) {
               const int& ai = au.blocks[i];
               const auto& arai = arSet[ai];
               const auto& blai = blSet[ai];
-              for (int j = i + 1; j < au.blocks.size(); ++j) {
+              for (int j = i + 1; j < ((int)au.blocks.size()); ++j) {
                 const int& bj = au.blocks[j];
                 const auto& arbj = arSet[bj];
                 const auto& blbj = blSet[bj];
@@ -1482,8 +1482,8 @@ bool SeqPair::CheckSymm(design& caseNL) {
 
 bool SeqPair::CheckAlign(design& caseNL) {
   for(auto align:caseNL.Align_blocks){
-    for (int i = 0; i < align.blocks.size() - 1; ++i) {
-      for (int j = i + 1; j < align.blocks.size(); ++j) {
+    for (int i = 0; i < ((int)align.blocks.size()) - 1; ++i) {
+      for (int j = i + 1; j < ((int)align.blocks.size()); ++j) {
         int first_it_pos, second_it_pos, first_it_neg, second_it_neg;
         first_it_pos = find(posPair.begin(), posPair.end(), align.blocks[i]) - posPair.begin();
         second_it_pos = find(posPair.begin(), posPair.end(), align.blocks[j]) - posPair.begin();
@@ -1522,8 +1522,8 @@ bool SeqPair::CheckAlign(design& caseNL) {
                 }
               }
               for(auto otheralign:caseNL.Align_blocks){
-                for (int i = 0; i < otheralign.blocks.size() - 1; ++i) {
-                  for (int j = i + 1; j < otheralign.blocks.size(); ++j) {
+                for (int i = 0; i < ((int)otheralign.blocks.size()) - 1; ++i) {
+                  for (int j = i + 1; j < ((int)otheralign.blocks.size()); ++j) {
                     if(otheralign.horizon){
                       if (a == otheralign.blocks[i] && b == otheralign.blocks[j] || a == otheralign.blocks[j] && b == otheralign.blocks[i]) return false;
                       // check other align pairs
@@ -1551,8 +1551,8 @@ bool SeqPair::CheckAlign(design& caseNL) {
                 }
               }
               for(auto otheralign:caseNL.Align_blocks){
-                for (int i = 0; i < otheralign.blocks.size() - 1; ++i) {
-                  for (int j = i + 1; j < otheralign.blocks.size(); ++j) {
+                for (int i = 0; i < ((int)otheralign.blocks.size()) - 1; ++i) {
+                  for (int j = i + 1; j < ((int)otheralign.blocks.size()); ++j) {
                     if(!otheralign.horizon){
                       if (a == otheralign.blocks[i] && b == otheralign.blocks[j] || a == otheralign.blocks[j] && b == otheralign.blocks[i]) return false;
                       // check other align pairs
